requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
python-multipart
cloudinary
pydantic
requests
anyio
starlette
typing-extensions
